services/verification/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
Pillow==12.1.1
numpy==1.26.4
pydantic==2.10.3
prometheus-client==0.21.0
